vrend/egl: ensure pre-signaled native fence is created before export
EGL_ANDROID_native_fence_sync expects that a context Flush() has
occured before an attempt to eglDupNativeFenceFDANDROID(). vrend is
missing that explicit flush for the sole pre-signaled fence. On Gallium
drivers, the dri frontend forces a flush during eglCreateSyncKHR, but that is
implementation-defined behavior that should not be relied upon.

Fixes: 313f1cc8879c ("vrend: add support for exporting egl fences")
Signed-off-by: Ryan Neph <ryanneph@google.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1168>