tgsi: simplify opcode declaration
The opcode IDs are not in sync with mesa, and we actually never need
specific opcode IDs, because we translate the shaders from text.
With that opcode TGSI_OPCODE_U642D becomes valid again (with a value
of 256 it was out of range for the eight bit value in tgsi_instruction.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1226>