shader: make sure to never exceed the immediate range
Make sure to always use an unsigned integer to compare the length of the immediate array.
Also be consistent with the use of MAX_IMMEDIATE everywhere.
Finally, give the compiler a hint on how likely are we to hit the failing case.

Signed-off-by: Corentin Noël <corentin.noel@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1231>