vrend: Read only as many bytes as were mapped
The width given in glMapBufferRange refers to machine units, i.e. the
format of the underlying buffer is of no consequence. Hence only
info->box->width bytes are mapped and can be copied to the IOV.

In guest mesa this is accounted for so we don't have to calculate the
send_size value.

Closes: https://gitlab.freedesktop.org/virgl/virglrenderer/-/issues/469
Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1233>