vrend: Check whether a texture view can be created before blitting
We don't want to try to create a texture view if this is not possible,
because if we do we will use a texture handle that is not backed by
anything in the blitting code.

Also fix the expectations since one more piglit passes.

Closes: https://gitlab.freedesktop.org/virgl/virglrenderer/-/issues/483

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1253>