vrend/video: Rename 'id' to 'gl_id' where appropriate
Resources and objects that are backed by a GL resource reference these
by the ID that is given out by glGen* and glCreate* calls. However, from
the naming it was not clear that these refer to GL resources, so rename
'ìd`to 'gl_id' to make it clearer what kind of ID we are dealing with.

With that we have also clearer what it means if that id is zero: it means
that the resource is not backed by a GL resource.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1276>