vrend: Check whether certain resources are backed by an OpenGL resources
It is possible to create a resource with PIPE_BIND_CUSTOM and this this
case they are only backed by host memory, but not by and OpenGL resource.
Trying to bind such a resource may later lead to crashes, because the host
OpenGL implementation will assume that data is provided by a user pointer
and will interpret the value that is supposed to be an offset into the GL
resource as a host pointer.
Make sure that we don't try to use a resource created with PIPE_BIND_CUSTOM
in these cases by checking it's gl_id which should be non-zero.

v2: add a patch to rename id to gl_id to make it clearer that we check for the
    existence of a backing by a GL resource (in discussion with tintou)

Closes: https://gitlab.freedesktop.org/virgl/virglrenderer/-/issues/494

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1276>