util: Enable list_assert in debugoptimized builds
See the reasoning in Mesa commit ed06b3dd4917a13d2f547a649f3689f3f7fa5ef2

In debugoptimized builds, DEBUG is not set (and neither is NDEBUG). The
intention of NDEBUG is to disable assertions. As such, list assertions should be
gated on !NDEBUG as opposed to on DEBUG.

Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1296>