vrend: Unref SO targets when SO object creation fails
When we set the streamout targets we add references to the targets that
are sucessfully bound. However, if an invalid resource object handle is
found the streamout object is destroyed. In order to avoid a memory leak
we have to unref the already bound SO targets again.

Closes: https://gitlab.freedesktop.org/virgl/virglrenderer/-/issues/508

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1301>