vrend: Fix GLSL version query
When we enter vrend_create_context for the first time the context
created for querying the supported version and also querying a
few GL limits has actually been destroyed, and for that reason
querying the GLSL version is also likely to fail and does so with
Qemu.
Therfore, query the GLSL version only after the first subcontext
has been created and made current.

Fixes: 7ca3ce97 (vrend: handle glsl version query failures as fatal)

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Part-of: <https://gitlab.freedesktop.org/virgl/virglrenderer/-/merge_requests/1312>